+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 
 - Add functionality to ignore SVGs based on a glob. ([#175], [#179])
 - Add option to customize Pull Request comments. ([#184])
+- Update the documentation for the Action's options. ([#185])
 
 ## [0.4.3] - 2020-04-10
 
@@ -116,3 +117,4 @@ Versioning].
 [#178]: https://github.com/ericcornelissen/svgo-action/pull/178
 [#179]: https://github.com/ericcornelissen/svgo-action/pull/179
 [#184]: https://github.com/ericcornelissen/svgo-action/pull/184
+[#185]: https://github.com/ericcornelissen/svgo-action/pull/185