+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning].
 ## [Unreleased]
 
 - Add functionality to ignore SVGs based on a glob. ([#175], [#179])
-- Add option to customize Pull Request comments. ([#184])
+- Add option to customize Pull Request comments. ([#184], [#186])
 - Update the documentation for the Action's options. ([#185])
 
 ## [0.4.3] - 2020-04-10
@@ -118,3 +118,4 @@ Versioning].
 [#179]: https://github.com/ericcornelissen/svgo-action/pull/179
 [#184]: https://github.com/ericcornelissen/svgo-action/pull/184
 [#185]: https://github.com/ericcornelissen/svgo-action/pull/185
+[#186]: https://github.com/ericcornelissen/svgo-action/pull/186